requirements.txt
crewai
torch
transformers
openai
pillow
numpy
pandas
python-dotenv
sentencepiece
accelerate
gradio
bitsandbytes